Implemented save fields for all item numbers

+++ src/main/java/com/itemconfiguration/service/MandatoryTranslationService.java
@@ -2,10 +2,12 @@ package com.itemconfiguration.service;
 
 import com.itemconfiguration.dao.MandatoryTranslationRepository;
 import com.itemconfiguration.domain.MandatoryTranslation;
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,12 +25,15 @@ public class MandatoryTranslationService {
 
 	@Transactional
 	public void deleteALl(List<MandatoryTranslation> mandatoryTranslations) {
-		this.mandatoryTranslationRepository.deleteByIds(getTranslationIds(mandatoryTranslations));
+		if (CollectionUtils.isNotEmpty(mandatoryTranslations)) {
+			this.mandatoryTranslationRepository.deleteByIds(getTranslationIds(mandatoryTranslations));
+		}
 	}
 
 	private List<Long> getTranslationIds(List<MandatoryTranslation> mandatoryTranslations) {
 		return mandatoryTranslations.stream()
 				.map(MandatoryTranslation::getId)
+				.filter(Objects::nonNull)
 				.collect(Collectors.toList());
 	}
